requirements.txt
flask==3.0.3
python-telegram-bot==21.0.1
flask-socketio==5.3.6
gunicorn==21.2.0